Adding PT-BR translation (#38)
* String translation to PT-BR

This is the initial implementation of the PT-BR localization to the BitD Foundry system.
Wherever possible, translations from the official Brazilian version by Redbox Editora were used.
Minor changes were made so that translations fit the UI better.

TO-DO:
- Contribute to additional, untranslated strings.
- Change some localization areas that may share strings on English but not in other languages.
- Change some structural/grammar choices that use translated strings but don't make sense in languages other than English

* Added PT-BR mention on README

* Removing temporary alternate system ID

This ID was originally added for local testing - it shouldn't have been committed

* Minor corrections

* Additional minor updates to PT-BR

* (PT-BR) Updated class name to match official translation

* (PT-BR) Updated XP terminology